--- src/rds-monitoring-mcp-server/awslabs/rds_monitoring_mcp_server/metrics_guide.py
@@ -1,38 +1,0 @@
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from loguru import logger
-from pathlib import Path
-
-
-def load_markdown_file(filename: str) -> str:
-    """Load a markdown file from the static/react directory.
-
-    Args:
-        filename (str): The name of the markdown file to load (e.g. 'basic-ui-setup.md')
-
-    Returns:
-        str: The content of the markdown file, or empty string if file not found
-    """
-    base_dir = Path(__file__).parent.parent
-    static_dir = base_dir / 'rds_monitoring_mcp_server/static'
-    file_path = static_dir / filename
-
-    if file_path.exists():
-        with open(file_path, 'r', encoding='utf-8') as f:
-            logger.info(f'Loading markdown file: {file_path}')
-            return f.read()
-    else:
-        logger.error('File not found: {file_path}')
-        return f'Warning: File not found: {file_path}'
