+++ src/rds-monitoring-mcp-server/awslabs/rds_monitoring_mcp_server/resources/db_cluster/__init__.py
@@ -1,0 +1,21 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This file is part of the awslabs namespace.
+# It is intentionally minimal to support PEP 420 namespace packages.
+
+from .cluster_discovery import get_cluster_details, get_cluster_overview
+from .list_cluster_metrics import list_cluster_metrics
+
+__all__ = ['get_cluster_details', 'get_cluster_overview', 'list_cluster_metrics']
